buildable project with header file
