+++ rr_endline_detection/src/endline_detection.cpp
@@ -1,0 +1,23 @@
+
+#include "endline_detection.hpp"
+
+EndlineCounter::EndlineCounter (ros::NodeHandle nh) : it(nh) 
+{
+	pub = nh.advertise<std_msgs::Bool>("endline",1);
+	this->state = BEGINNING;
+}
+
+// void EndlineCounter::img_callback(const sensor_msgs::ImagePtr& msg) {
+	
+// 	// process image
+	
+	
+// 	// publish 
+// 	// if(this->state & PAST_END) {
+// 	// 	this->pub.publish(true);
+// 	// } else {
+// 	// 	this->pub.publish(false);
+// 	// }
+// }
+
+void EndlineCounter::find_el_state(){}
