+++ rr_endline_detection/src/main.cpp
@@ -7,81 +7,28 @@
 #include <image_transport/image_transport.h>
 #include <cv_bridge/cv_bridge.h>
 #include <sensor_msgs/image_encodings.h>
-
-using namespace cv;
-static volatile int state;
-
-enum states {
-	BEGINNING = 0x00,
-	START_LINE = 0x03,
-	MIDDLE = 0x04,
-	END_LINE = 0x07,
-	PAST_END = 0x08
-};
+#include <opencv2/opencv.hpp>
+#include "endline_detection.hpp"
+#include <std_msgs/Bool.h>
 
 
-class EndlineCounter {
-	private :
-		ros::NodeHandle nh;
-		
-		image_transport::ImageTransport it;
-		image_transport::Subscriber sub;
-		
-		cv_bridge::CvImagePtr imgptr;
-		cv::Mat img;
-		
-		ros::Publisher pub;
-	
-	
-	public :
-		int state;
-		
-		// EndlineCounter (ros::NodeHandle nh) : it(nh) 
-		// {
-		// 	sub = it.subscribe("tl_camera/image",1, &EndlineCounter::img_callback);
-		// 	nh.advertise<std_msgs::Bool>("endline",1);
-		// 	state = BEGINNING;
-		// }
-		int get_state() {
-			return state;
-		}
-		
-};
+void imageCallback (const sensor_msgs::ImageConstPtr& msg) {
 
-// EndlineCounter::img_callback(const sensor_msgs::ImagePtr& msg) {
-// 	pub = nh.advertise<std_msgs::Bool>("endline", 1);
-// 	imgptr = cv_bridge::toCvCopy(msg, "bgr8")->image); 
-	
-// 	//process image
-	
-	
-// 	//publish 
-// 	if(state & ENDLINE_PAST) {
-// 		pub.publish(true);
-// 	} else {
-// 		pub.publish(false);
-// 	}
-// }
-
-// void EndlineCounter::find_el_state() {
-	
-// }
-
+}
 
 int main(int argc, char** argv)
 {
-	
 	ros::init(argc, argv, "endline_detection");
-	std::cout << "Initializing endline_detection\n\r";
+	//  "Initializing endline_detection\n\r";
 	
-	//nodehandler.param ?
+    std::string camera_source;
+    ros::NodeHandle nh;
+	image_transport::ImageTransport it(nh);
+    //nh.param<std::string>("Camera_Source_Topic", camera_source, "/usb_cam/image_raw");
+    // EndlineCounter ec(nh);
+    image_transport::Subscriber sub = it.subscribe("camera_source",1, imageCallback);
 	
-	
-	while(ros::ok)
-	{
-	
-		
-	}
+	ros::spin();
 
     return 0;
 }
